// desktop/src/open-api/model/filterOperation.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type FilterOperation = 'CONTAINS' | 'EQUALS' | 'GREATER_THAN' | 'LESS_THAN' | 'BETWEEN' | 'WITHIN_CURRENT_MONTH' | '';

export const FilterOperation = {
    Contains: 'CONTAINS' as FilterOperation,
    Equals: 'EQUALS' as FilterOperation,
    GreaterThan: 'GREATER_THAN' as FilterOperation,
    LessThan: 'LESS_THAN' as FilterOperation,
    Between: 'BETWEEN' as FilterOperation,
    WithinCurrentMonth: 'WITHIN_CURRENT_MONTH' as FilterOperation,
    Empty: '' as FilterOperation
};